+++ src/client/kotlin/joshxviii/plantz/PazScreens.kt
@@ -1,9 +1,7 @@
 package joshxviii.plantz
 
-import joshxviii.plantz.block.entity.MailboxManager
 import joshxviii.plantz.gui.MailboxScreen
 import joshxviii.plantz.gui.TimeMachineScreen
-import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientLevelEvents
 import net.minecraft.client.gui.screens.MenuScreens
 
 object PazScreens {