+++ src/main/kotlin/joshxviii/plantz/PazMain.kt
@@ -1,9 +1,15 @@
 package joshxviii.plantz
 
+import joshxviii.plantz.block.entity.MailboxBlockEntity
+import joshxviii.plantz.block.entity.MailboxManager
+import joshxviii.plantz.networking.SendMailResponsePayload
 import joshxviii.plantz.raid.getZombieRaids
 import net.fabricmc.api.ModInitializer
+import net.fabricmc.fabric.api.event.lifecycle.v1.ServerBlockEntityEvents
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerTickEvents
+import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking
+import net.minecraft.network.chat.Component
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 
@@ -19,6 +25,18 @@ object PazMain : ModInitializer {
 		ServerLifecycleEvents.SERVER_STARTING.register { PazConfig.load() }
 
 		ServerTickEvents.END_LEVEL_TICK.register { it.getZombieRaids().tick(it) }
+
+		// mailbox managing
+		ServerBlockEntityEvents.BLOCK_ENTITY_LOAD.register { blockEntity, level ->
+			(blockEntity as? MailboxBlockEntity)?.let {
+				MailboxManager.registerMailbox(level, it)
+			}
+		}
+		ServerBlockEntityEvents.BLOCK_ENTITY_UNLOAD.register { blockEntity, level ->
+			(blockEntity as? MailboxBlockEntity)?.let {
+				MailboxManager.unregisterMailbox(level, blockEntity.blockPos)
+			}
+		}
 
 		PazServerParticles.initialize()
 		PazBlocks.initialize()