+++ src/main/kotlin/joshxviii/plantz/PazNetwork.kt
@@ -1,7 +1,8 @@
 package joshxviii.plantz
 
-import joshxviii.plantz.networking.SendMailPayload
-import joshxviii.plantz.networking.SendMailPayload.Companion.handleSendMailPacket
+import joshxviii.plantz.network.MailboxListResponsePayload
+import joshxviii.plantz.networking.SendMailRequestPayload
+import joshxviii.plantz.networking.SendMailRequestPayload.Companion.handleSendMailPacket
 import joshxviii.plantz.networking.SendMailResponsePayload
 import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking
@@ -10,12 +11,13 @@ object PazNetwork {
 
     fun initialize() {
         // Register payloads
-        PayloadTypeRegistry.clientboundPlay().register(SendMailPayload.ID, SendMailPayload.STREAM_CODEC)
-        PayloadTypeRegistry.serverboundPlay().register(SendMailPayload.ID, SendMailPayload.STREAM_CODEC)
+        PayloadTypeRegistry.serverboundPlay().register(SendMailRequestPayload.ID, SendMailRequestPayload.STREAM_CODEC)
 
+        PayloadTypeRegistry.clientboundPlay().register(SendMailRequestPayload.ID, SendMailRequestPayload.STREAM_CODEC)
         PayloadTypeRegistry.clientboundPlay().register(SendMailResponsePayload.ID, SendMailResponsePayload.STREAM_CODEC)
+        PayloadTypeRegistry.clientboundPlay().register(MailboxListResponsePayload.ID, MailboxListResponsePayload.STREAM_CODEC)
 
         // Register server receiver
-        ServerPlayNetworking.registerGlobalReceiver(SendMailPayload.ID, ::handleSendMailPacket)
+        ServerPlayNetworking.registerGlobalReceiver(SendMailRequestPayload.ID, ::handleSendMailPacket)
     }
 }