+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 - Disco zombie can now only spawn backup when there are < 3 backup dances nearby rather than < 4
 - Updated list of projectiles that wallnut can deflect.
 - Increased the amount of healing sun gives to plants.
+- Mailboxes have been updated.
+  - All players can now access mailboxes from any loaded chunk within the same dimension.
+  - The address list is now sorted by distance from the current mailbox.
 
 # Changelog 1.4
 - Added Doom-Shroom